+++ gameAssets/GameStart/cardDeck.json
@@ -4,14 +4,14 @@
     { "type": 1, "count": 0 },
     { "type": 2, "count": 0 },
     { "type": 3, "count": 0 },
-    { "type": 4, "count": 20 },
+    { "type": 4, "count": 0 },
     { "type": 5, "count": 0 },
     { "type": 6, "count": 0 },
     { "type": 7, "count": 0 },
     { "type": 8, "count": 0 },
     { "type": 9, "count": 0 },
     { "type": 10, "count": 0 },
-    { "type": 11, "count": 0 },
+    { "type": 11, "count": 20 },
     { "type": 12, "count": 20 },
     { "type": 13, "count": 1 },
     { "type": 14, "count": 2 },
