--- src/utils/card/lotto copy.js
@@ -1,60 +1,0 @@
-import { CARD_TYPE } from '../../constants/card.enum.js';
-import HANDLER_IDS from '../../constants/handlerIds.js';
-import { CHARACTER_STATE_TYPE } from '../../constants/user.enum.js';
-import playerList from '../../model/player/playerList.class.js';
-import roomList from '../../model/room/roomList.class.js';
-import createFailCode from '../../utils/response/createFailCode.js';
-import { createResponse } from '../../utils/response/createResponse.js';
-
-const lotto = ({ socket, cardType, targetUserId }) => {
-  const user = playerList.getPlayer(socket.id);
-  
-  const room = roomList.getRoom(user.currentRoomId);
-  const targetUser = playerList.getPlayer(targetUserId.low);
-  console.log(user);
-  //3장 드로우및 핸드수 증가.
-  user.characterData.handCards.push(room.cardDraw(3));
-  console.log(user.characterData.handCards);
-  user.increaseHandCardsCountParam(3);
-
-  const inGameUsers = Array.from(room.getAllPlayers().values());
-
-  const S2CUseCardNotification = {
-    cardType: cardType,
-    userId: user.id,
-    targetUserId: targetUser.id,
-  };
-
-  inGameUsers.forEach((player) => {
-    const gamePacket = { useCardNotification: S2CUseCardNotification };
-s
-    const useCardNotification = createResponse(
-      HANDLER_IDS.USE_CARD_NOTIFICATION,
-      player.socket.version,
-      player.socket.sequence,
-      gamePacket,
-    );
-
-    player.socket.write(useCardNotification);
-
-    const S2CUserUpdateNotification = { user: player.getAllUsersData() };
-
-    const updatePacket = { userUpdateNotification: S2CUserUpdateNotification };
-
-    const userUpdateNotification = createResponse(
-      HANDLER_IDS.USER_UPDATE_NOTIFICATION,
-      player.socket.version,
-      player.socket.sequence,
-      updatePacket,
-    );
-
-    player.socket.write(userUpdateNotification);
-  });
-
-  return {
-    success: true,
-    failCode: createFailCode(0),
-  };
-};
-
-export default lotto;
